pyproject.toml
[project]
name = "ximrato-server"
version = "0.1.0"
description = "FastAPI backend for ximrato-app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "pydantic[email]>=2.12.5",
    "bcrypt>=5.0.0",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]
